+++ source/voxelman/server/clientinfo.d
@@ -15,7 +15,7 @@ struct ClientInfo
 	string name;
 	vec3 pos = START_POS;
 	vec2 heading = vec2(0, 0);
-	size_t viewRadius = VIEW_RADIUS;
+	int viewRadius = VIEW_RADIUS;
 	ChunkRange visibleRegion;
 
 	ENetPeer* peer;
